/0 예외처리 수정 Integer -> Double

+++ Calculation/src/main/java/com/programmers/java/calculation/calculate/CalculateBasicImpl.java
@@ -26,7 +26,7 @@ public class CalculateBasicImpl implements Calculate {
                 if (Objects.equals(input.get(i), "*")) {
                     mulDivResult = nums.get(lastIndex) * Double.parseDouble(input.get(i+1));
                 } else {
-                    if (Integer.parseInt(input.get(i+1)) == 0) {
+                    if (Double.parseDouble(input.get(i+1)) == 0) {
                         return null;
                     }
                     mulDivResult = nums.get(lastIndex) / Double.parseDouble(input.get(i+1));
